로그아웃 후 삭제하려고 하면 401뜨고, 로그인 후 다른 사용자의 글 삭제 하려면 403. 로그인 후 자기글 삭제 200

+++ src/main/java/com/example/prj1be/controller/BoardController.java
@@ -48,7 +48,18 @@ public class BoardController {
     }
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id) {
+    public ResponseEntity remove(@PathVariable Integer id,
+                                 @SessionAttribute(value = "login", required = false) Member login) {
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); //401. 로그인이 안됨
+        }
+
+        // 접근 권한 없으면 403 응답해줌
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); //403. 누군지는 알지만 권한이 없음
+        }
+
         if (service.remove(id)) {
             return ResponseEntity.ok().build();
         } else {
